requirements.txt
django
django-autocomplete
